+++ changelog.md
@@ -7,8 +7,7 @@ compromise uses semver, and pushes to npm and github frequently
 While all _Major_ releases should be reviewed, our only _large_ releases are **v6** in 2016 **v12** in 2019 and **v14** in 2022. Others have been mostly incremental.
 
 
-
-#### [unreleased]
+#### 14.17.0 [Sept 2026]
 - **[new]** - split japanese/chinese sentences on 。！？ without whitespace, respecting 「」（）brackets
 - **[fix]** - sentence-split for korean, and cjk text with ascii periods
 - **[fix]** - fullwidth ？ in .questions()
@@ -16,6 +15,7 @@ While all _Major_ releases should be reviewed, our only _large_ releases are **v
 - **[new]** - normalize fullwidth ascii - 'Ｈｅｌｌｏ ２０２４'
 - **[fix]** - '“Yes!” said Tom. “No!” said Ann.' merged into one sentence
 - **[fix]** - replace all non-breaking spaces, not only the first
+- **[fix]** - unicode non-breaking spaces in input #1220
 
 #### 14.16.0 [July 2026]
 - **[new]** - apply redaction to #Organization, #Address, #Url etc tags in .redact() method